feat: working on tests
