+++ sturdy/abi/SturdySiloStrategy.json
@@ -31,6 +31,19 @@
         "type": "constructor"
     },
     {
+        "name": "pricePerShare",
+        "type": "function",
+        "inputs": [],
+        "outputs": [
+            {
+                "name": "",
+                "type": "uint256",
+                "internalType": "uint256"
+            }
+        ],
+        "stateMutability": "view"
+    },
+    {
         "stateMutability": "nonpayable",
         "type": "fallback"
     },
