+++ tests/integration/validator/test_integration_validator.py
@@ -1,78 +1,260 @@
 import copy
+import os
 import unittest
 from unittest import IsolatedAsyncioTestCase
 
 import numpy as np
+import torch
+from dotenv import load_dotenv
 
 from neurons.validator import Validator
+from sturdy.constants import QUERY_TIMEOUT
+from sturdy.mock import MockDendrite
 from sturdy.pools import generate_challenge_data
+from sturdy.protocol import REQUEST_TYPES, AllocateAssets, AllocationsDict
 from sturdy.validator.forward import query_and_score_miners
-from sturdy.validator.simulator import Simulator
+from sturdy.validator.reward import get_rewards
 
+load_dotenv()
+os.environ["WEB_PROVIDER_URL"] = "http://127.0.0.1:8545"
+WEB3_PROVIDER_URL = os.getenv("WEB3_PROVIDER_URL")
 
+# TODO: more comprehensive integration testing?
 class TestValidator(IsolatedAsyncioTestCase):
     maxDiff = 4000
 
     @classmethod
     def setUpClass(cls) -> None:
-        # dont log this in wandb
+        np.random.seed(69)  # noqa: NPY002
         config = {
             "mock": True,
             "wandb": {"off": True},
             "mock_n": 16,
             "neuron": {"dont_save_events": True},
-            "netuid": 69,
         }
-        cls.validator = Validator(config=config)
-        # simulator with preset seed
-        cls.validator.simulator = Simulator(seed=69)
 
-        assets_and_pools = generate_challenge_data(np.random.RandomState(seed=420))  # type: ignore[]
+        cls.validator = Validator(config=config)
+        assert cls.validator.w3.is_connected()
+
+        cls.validator.w3.provider.make_request(
+            "hardhat_reset",  # type: ignore[]
+            [
+                {
+                    "forking": {
+                        "jsonRpcUrl": WEB3_PROVIDER_URL,
+                        "blockNumber": 21080765,
+                    },
+                },
+            ],
+        )
+
+        generated_data = generate_challenge_data(cls.validator.w3, np.random.RandomState(seed=420))
+        assets_and_pools = generated_data["assets_and_pools"]
+
+        cls.contract_addresses: list[str] = list(assets_and_pools["pools"].keys())  # type: ignore[]
+
+        assets_and_pools["pools"][cls.contract_addresses[0]].borrow_amount = int(75e18)
+        assets_and_pools["pools"][cls.contract_addresses[1]].borrow_amount = int(50e18)
+        assets_and_pools["pools"][cls.contract_addresses[2]].borrow_amount = int(85e18)
+        assets_and_pools["pools"][cls.contract_addresses[3]].borrow_amount = int(25e18)
+        assets_and_pools["pools"][cls.contract_addresses[4]].borrow_amount = int(90e18)
+        assets_and_pools["pools"][cls.contract_addresses[5]].borrow_amount = int(25e18)
+        assets_and_pools["pools"][cls.contract_addresses[6]].borrow_amount = int(25e18)
+        assets_and_pools["pools"][cls.contract_addresses[7]].borrow_amount = int(40e18)
+        assets_and_pools["pools"][cls.contract_addresses[8]].borrow_amount = int(45e18)
+        assets_and_pools["pools"][cls.contract_addresses[9]].borrow_amount = int(80e18)
 
         cls.assets_and_pools = {
             "pools": assets_and_pools["pools"],
             "total_assets": int(1000e18),
         }
 
-        cls.contract_addresses = list(assets_and_pools["pools"].keys())  # type: ignore[]
-
-        cls.allocations = {
-            cls.contract_addresses[0]: 100e18,
-            cls.contract_addresses[1]: 100e18,
-            cls.contract_addresses[2]: 200e18,
-            cls.contract_addresses[3]: 50e18,
-            cls.contract_addresses[4]: 200e18,
-            cls.contract_addresses[5]: 25e18,
-            cls.contract_addresses[6]: 25e18,
-            cls.contract_addresses[7]: 50e18,
-            cls.contract_addresses[8]: 50e18,
-            cls.contract_addresses[9]: 200e18,
+        cls.allocations: AllocationsDict = {
+            cls.contract_addresses[0]: int(100e18),
+            cls.contract_addresses[1]: int(100e18),
+            cls.contract_addresses[2]: int(200e18),
+            cls.contract_addresses[3]: int(50e18),
+            cls.contract_addresses[4]: int(200e18),
+            cls.contract_addresses[5]: int(45e18),
+            cls.contract_addresses[6]: int(45e18),
+            cls.contract_addresses[7]: int(50e18),
+            cls.contract_addresses[8]: int(50e18),
+            cls.contract_addresses[9]: int(160e18),
         }
 
+    async def test_get_rewards(self) -> None:
+        print("----==== test_get_rewards ====----")
+
+        assets_and_pools = copy.deepcopy(self.assets_and_pools)
+        allocations = copy.deepcopy(self.allocations)
+
+        validator = self.validator
+
+        validator.simulator.init_data(
+            init_assets_and_pools=copy.deepcopy(assets_and_pools),
+            init_allocations=copy.deepcopy(allocations),
+        )
+
+        active_uids = [str(uid) for uid in range(validator.metagraph.n.item()) if validator.metagraph.axons[uid].is_serving]  # type: ignore[]
+
+        active_axons = [validator.metagraph.axons[int(uid)] for uid in active_uids]
+
+        synapse = AllocateAssets(
+            request_type=REQUEST_TYPES.SYNTHETIC,
+            assets_and_pools=copy.deepcopy(assets_and_pools),
+            allocations=copy.deepcopy(allocations),
+        )
+
+        validator.dendrite = MockDendrite(wallet=validator.wallet, custom_allocs=True)
+        responses = await validator.dendrite(
+            # Send the query to selected miner axons in the network.
+            axons=active_axons,
+            # Construct a dummy query. This simply contains a single integer.
+            synapse=synapse,
+            deserialize=False,
+            timeout=QUERY_TIMEOUT,
+        )
+
+        for response in responses:
+            self.assertEqual(response.assets_and_pools, assets_and_pools)
+            self.assertLessEqual(sum(response.allocations.values()), assets_and_pools["total_assets"])
+
+        rewards, allocs = get_rewards(
+            validator,
+            validator.step,
+            active_uids,
+            responses=responses,
+            assets_and_pools=assets_and_pools,
+        )
+
+        print(f"allocs: {allocs}")
+
+        rewards_dict = {active_uids[k]: v for k, v in enumerate(list(rewards))}
+        sorted_rewards = dict(sorted(rewards_dict.items(), key=lambda item: item[1], reverse=True))  # type: ignore[]
+
+        print(f"sorted rewards: {sorted_rewards}")
+
+        # rewards should not all be the same
+        to_compare = torch.empty(rewards.shape)
+        torch.fill(to_compare, rewards[0])
+        self.assertFalse(torch.equal(rewards, to_compare))
+
+    async def test_get_rewards_punish(self) -> None:
+        print("----==== test_get_rewards_punish ====----")
+        validator = self.validator
+        assets_and_pools = copy.deepcopy(self.assets_and_pools)
+
+        allocations = copy.deepcopy(self.allocations)
+        # increase one of the allocations by +10000  -> clearly this means the miner is cheating!!!
+        allocations[self.contract_addresses[0]] += int(10000e18)
+
+        validator.simulator.reset()
+        validator.simulator.init_data(
+            init_assets_and_pools=copy.deepcopy(assets_and_pools),
+            init_allocations=copy.deepcopy(allocations),
+        )
+
+        active_uids = [str(uid) for uid in range(validator.metagraph.n.item()) if validator.metagraph.axons[uid].is_serving]  # type: ignore[]
+
+        active_axons = [validator.metagraph.axons[int(uid)] for uid in active_uids]
+
+        synapse = AllocateAssets(
+            request_type=REQUEST_TYPES.SYNTHETIC,
+            assets_and_pools=copy.deepcopy(assets_and_pools),
+            allocations=copy.deepcopy(allocations),
+        )
+
+        validator.dendrite = MockDendrite(wallet=validator.wallet)
+        responses = await validator.dendrite(
+            # Send the query to selected miner axons in the network.
+            axons=active_axons,
+            # Construct a dummy query. This simply contains a single integer.
+            synapse=synapse,
+            deserialize=False,
+            timeout=QUERY_TIMEOUT,
+        )
+
+        for response in responses:
+            self.assertEqual(response.assets_and_pools, assets_and_pools)
+            self.assertEqual(response.allocations, allocations)
+
+        rewards, allocs = get_rewards(
+            validator,
+            validator.step,
+            active_uids,
+            responses=responses,
+            assets_and_pools=assets_and_pools,
+        )
+
+        for allocInfo in allocs.values():
+            self.assertEqual(allocInfo["apy"], 0)
+
+        # rewards should all be the same (0)
+        self.assertEqual(all(rewards), 0)
+
+        rewards_dict = dict(enumerate(list(rewards)))
+        sorted_rewards = dict(sorted(rewards_dict.items(), key=lambda item: item[1], reverse=True))  # type: ignore[]
+
+        print(f"sorted rewards: {sorted_rewards}")
+
+        assets_and_pools = copy.deepcopy(self.assets_and_pools)
+
+        allocations = copy.deepcopy(self.allocations)
+        # set one of the allocations to be negative! This should not be allowed!
+        allocations[self.contract_addresses[0]] = -1
+
+        validator.simulator.reset()
+        validator.simulator.init_data(
+            init_assets_and_pools=copy.deepcopy(assets_and_pools),
+            init_allocations=copy.deepcopy(allocations),
+        )
+
+        active_uids = [str(uid) for uid in range(validator.metagraph.n.item()) if validator.metagraph.axons[uid].is_serving]  # type: ignore[]
+
+        active_axons = [validator.metagraph.axons[int(uid)] for uid in active_uids]
+
+        synapse = AllocateAssets(
+            request_type=REQUEST_TYPES.SYNTHETIC,
+            assets_and_pools=copy.deepcopy(assets_and_pools),
+            allocations=copy.deepcopy(allocations),
+        )
+
+        validator.dendrite = MockDendrite(wallet=validator.wallet)
+        responses = await validator.dendrite(
+            # Send the query to selected miner axons in the network.
+            axons=active_axons,
+            # Construct a dummy query. This simply contains a single integer.
+            synapse=synapse,
+            deserialize=False,
+            timeout=QUERY_TIMEOUT,
+        )
+
+        for response in responses:
+            self.assertEqual(response.assets_and_pools, assets_and_pools)
+            self.assertEqual(response.allocations, allocations)
+
+        rewards, allocs = get_rewards(
+            validator,
+            validator.step,
+            active_uids,
+            responses=responses,
+            assets_and_pools=assets_and_pools,
+        )
+
+        for allocInfo in allocs.values():
+            self.assertEqual(allocInfo["apy"], 0)
+
+        # rewards should all be the same (0)
+        self.assertEqual(all(rewards), 0)
+
+        rewards_dict = dict(enumerate(list(rewards)))
+        sorted_rewards = dict(sorted(rewards_dict.items(), key=lambda item: item[1], reverse=True))  # type: ignore[]
+
+        print(f"sorted rewards: {sorted_rewards}")
+
+
     async def test_query_and_score_miners(self) -> None:
-        # use simulator generated assets and pools
-        await query_and_score_miners(self.validator)
-        self.assertIsNotNone(self.validator.simulator.assets_and_pools)
-        self.assertIsNotNone(self.validator.simulator.allocations)
-        self.maxDiff = None
-
-        # use user-defined generated assets and pools
-        simulator_copy = copy.deepcopy(self.validator.simulator)
-        await query_and_score_miners(
-            self.validator,
-            assets_and_pools=copy.deepcopy(self.assets_and_pools),
-        )
-        simulator_copy.initialize()
-        simulator_copy.init_data(
-            init_assets_and_pools=copy.deepcopy(self.assets_and_pools),
-        )
-        simulator_copy.update_reserves_with_allocs()
-        # TODO: update these tests - low priority
-        # assets_pools_should_be = simulator_copy.assets_and_pools
-
-        # assets_pools2 = self.validator.simulator.assets_and_pools
-        # self.assertEqual(assets_pools2, assets_pools_should_be)
-        # self.assertIsNotNone(self.validator.simulator.allocations)
+        await query_and_score_miners(self.validator, assets_and_pools=self.assets_and_pools)
 
     async def test_forward(self) -> None:
         await self.validator.forward()
